# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Dreamshelix India EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [payments]
      summary: Get withdrawal history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: requestWithdrawal
      tags: [payments]
      summary: Request a withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/withdrawals:
    get:
      operationId: listAdminWithdrawals
      tags: [admin]
      summary: List all withdrawals (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{withdrawalId}/approve:
    put:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve withdrawal (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{withdrawalId}/reject:
    put:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject withdrawal (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/referrals:
    get:
      operationId: listAdminReferrals
      tags: [admin]
      summary: List all referrals (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All referrals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferralRecord"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        referralCode:
          type: string
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, admin]
        referralCode:
          type: string
        walletBalance:
          type: number
        totalEarnings:
          type: number
        isActive:
          type: boolean
        avatar:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - referralCode
        - walletBalance
        - totalEarnings
        - isActive
        - createdAt
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, admin]
        isActive:
          type: boolean
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        bankDetails:
          type: string
        upiId:
          type: string
        notes:
          type: string
        processedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - status
        - createdAt
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank_transfer, upi]
        bankDetails:
          type: string
        upiId:
          type: string
      required:
        - amount
        - method
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        newUsersThisMonth:
          type: integer
        totalCourses:
          type: integer
        totalEnrollments:
          type: integer
        totalRevenue:
          type: number
        revenueThisMonth:
          type: number
        totalWithdrawalsPending:
          type: integer
        totalWithdrawalAmount:
          type: number
        revenueChart:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        enrollmentChart:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        topCourses:
          type: array
          items:
            $ref: "#/components/schemas/TopCourse"
      required:
        - totalUsers
        - newUsersThisMonth
        - totalCourses
        - totalEnrollments
        - totalRevenue
        - revenueThisMonth
        - totalWithdrawalsPending
        - totalWithdrawalAmount
        - revenueChart
        - enrollmentChart
        - topCourses
    ChartDataPoint:
      type: object
      properties:
        label:
          type: string
        value:
          type: number
      required:
        - label
        - value
    TopCourse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        enrollments:
          type: integer
        revenue:
          type: number
      required:
        - id
        - title
        - enrollments
        - revenue
    ReferralRecord:
      type: object
      properties:
        id:
          type: integer
        referrerId:
          type: integer
        referredId:
          type: integer
        referrerName:
          type: string
        referredName:
          type: string
        level:
          type: string
          enum: [direct, indirect]
        commissionAmount:
          type: number
        status:
          type: string
          enum: [pending, credited]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - referrerId
        - referredId
        - level
        - commissionAmount
        - status
        - createdAt
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - totalPages