# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dreamshelix India EdTech Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
  role: UserRole;
  referralCode: string;
  walletBalance: number;
  totalEarnings: number;
  isActive: boolean;
  avatar?: string;
  createdAt: Date;
}